🔄 Refactor : detach path relatives
